+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## v1.0.9
+- Added `OS`, `Process`, `ObjectInteract` class
+- New features `$throw`, `$new`, `$tryAndCatch`, `$let`, `$get`
 ## v1.0.7
 - Added `CHANGELOG.md`
 - Added Environment Manager